[#231] feat : 노션 페이지 적용
