+++ app/settings/_components/Service/index.tsx
@@ -1,52 +1,20 @@
 "use client";
-import Link from "next/link";
-import { ExtendedRecordMap } from "notion-types";
-import { NotionRenderer } from "react-notion-x";
-import { useEffect, useState } from "react";
+
 import pageInfoMap, { PageInfo } from "@/app/settings/_components/Service/PageInfo";
 import * as styles from "@/app/settings/_components/Service/style.css";
 import CloseIcon from "@/public/icons/close.svg";
+import "katex/dist/katex.min.css";
 import { useRouter } from "next/navigation";
+import { ExtendedRecordMap } from "notion-types";
+import "prismjs/themes/prism-tomorrow.css";
+import * as React from "react";
+import "react-notion-x/src/styles.css";
+import NotionPage from "./_components/NotionPage";
 
-const notionDomain = "https://hissing-route-336.notion.site";
-
-interface NotionResultProps {
-  rootPageId: string;
-  rootDomain: string;
-  previewImagesEnabled: boolean;
-  recordMap: ExtendedRecordMap;
-}
 interface ServiceProps {
   recordMap: ExtendedRecordMap;
   pageType: "faq" | "ask" | "notice";
 }
-
-const NotionResult: React.FC<NotionResultProps> = ({ rootPageId, previewImagesEnabled, recordMap }) => {
-  const [loadedRecordMap, setLoadedRecordMap] = useState<ExtendedRecordMap | null>(null);
-
-  useEffect(() => {
-    setLoadedRecordMap(recordMap);
-  }, [recordMap]);
-
-  if (!loadedRecordMap) {
-    return null;
-  }
-
-  return (
-    <>
-      <NotionRenderer
-        recordMap={loadedRecordMap}
-        rootDomain={notionDomain}
-        rootPageId={rootPageId}
-        fullPage={false}
-        previewImages={previewImagesEnabled}
-        components={{
-          nextLink: Link,
-        }}
-      />
-    </>
-  );
-};
 
 const Service: React.FC<ServiceProps> = ({ pageType, recordMap }) => {
   const pageInfo: PageInfo = pageInfoMap[pageType];
@@ -61,7 +29,7 @@ const Service: React.FC<ServiceProps> = ({ pageType, recordMap }) => {
         </div>
       </header>
       <main className={styles.mainWrapper}>
-        <NotionResult recordMap={recordMap} previewImagesEnabled={false} rootDomain={""} rootPageId={pageInfo.pageId} />
+        <NotionPage recordMap={recordMap} previewImagesEnabled={false} rootDomain={""} rootPageId={pageInfo.pageId} />
       </main>
     </>
   );
